requirements.txt
mysql-connector-python==8.0.33
matplotlib==3.7.2
ttkbootstrap==1.11.0
